Prevent state setting on unmounted component
It's a common problem in React that component state is set even though the component got already unmounted (e.g. due to navigating away with React Router).

With this solution, actually not the data fetching is aborted -- which could be achieved with Axios Cancellation -- but the state transition is not performed anymore for the unmounted component. Since Axios Cancellation has not the best API in my eyes, this boolean flag to prevent setting state does the job as well.